+++ test/RunnableFactory.test.ts
@@ -45,7 +45,9 @@ describe(path.basename(__filename), function () {
         {},
         {},
         {},
+        {},
         1,
+        false,
         {},
         new DefaultSpawner(),
       );
@@ -86,7 +88,9 @@ describe(path.basename(__filename), function () {
         {},
         {},
         {},
+        {},
         1,
+        false,
         {},
         new DefaultSpawner(),
       );