+++ test/framework/Catch2Runnable.test.ts
@@ -27,6 +27,7 @@ describe(pathlib.basename(__filename), function () {
     {},
     {},
     1,
+    false,
     {},
     new DefaultSpawner(),
   );